+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to **dot-diver** will be documented here. Inspired by [keep 
 
 ## Unreleased
 
+- Fixed getByPath throws if accessing a null value
+
 ## [1.0.4](https://github.com/clickbar/dot-diver/tree/1.0.4) (2023-11-07)
 
 - Fixed wrong type transformation via vite-dts-plugin (see #15)